Update Gitpod before script (#20)
Projects have suddenly having issues setting up in Gitpod. They're starting with "Operation not permitted" errors when chown-ing the pre-configured files from the Docker image.

The cause of this new behavior appears to be because the container now has creates a /home/gitpod/ directory prior to the before task executing, which wasn't the case before. The cp -R command behaves differently depending on if the destination folder exists or not. When the folder does exist the source directory is added as a sub-folder instead of adding the contents directly in the destination folder. The previous script started to create a sub-folder which was breaking the chown command.

The before task has been updated to:

- copy contents of student to gitpod (creating sub-folder)
- move the contents of the student sub-folder to it's parent gitpod folder.
- source ~/.bashrc to ensure Bash configuration is loaded before bin/setup

Additionally, the DATABASE_URL has been unset to allow for SQLite3 configuration (related to #11)