Deactivate Turbo initially (#29)
With Turbo enabled by default, form submissions and link navigation that would normally cause a "No Route matches" error, instead simply appear to refresh the page. While this is the expected behavior for Turbo (only updating the page when successful responses are sent) this is an unexpected behavior for beginners.

This commit deactivates Turbo initially, so that students in AppDev I/Phase 1 content are not confused by the described behavior. This also allows for more advanced students (AppDev II/Phase II) to re-activate Turbo if/when they need it.