Add alias for rake grade and reset_token (#32)
Per the linked grade_runner issue, we discussed there adding the CLI grade directly to the gem. However, this didn't seem totally straightforward, and, on that ticket we discussed that since we are requiring Codespaces (and Gitpod as a backup), and we have totally removed any "local setup" instructions from the content, then a simple alias in the Dockerfile is acceptable in the immediate short term.