Enable SSL enforcement on Production (#40)
In the past Heroku required paid dynos for auto SSL certs management, which is probably why we had it off. 

Render does not charge for this: https://docs.render.com/tls

Fly.io is also free for the first 10: https://fly.io/docs/about/pricing/#managed-ssl-certificates; and at only $0.10/mo per additional, this is likely still below the $5 "free limit"; in any case, we plan to promote Fly.io as a professional / long-term / paid portfolio option, and primarily stick with Render for deployment.